# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ascend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createUserProfile
      tags: [users]
      summary: Create or update user profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: resetUserProfile
      tags: [users]
      summary: Reset (delete) the current user's profile and all their data
      responses:
        "204":
          description: Reset complete
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /plans/current:
    get:
      operationId: getCurrentPlan
      tags: [plans]
      summary: Get current personalized plan
      responses:
        "200":
          description: Current plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: generatePlan
      tags: [plans]
      summary: Generate a personalized plan based on user profile
      responses:
        "201":
          description: Generated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
...
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List nightly journal entries
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Submit a nightly journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/today:
    get:
      operationId: getTodayJournalEntry
      tags: [journal]
      summary: Get today's journal entry
      responses:
        "200":
          description: Today's journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /reviews/today:
    get:
      operationId: getTodayReview
      tags: [reviews]
      summary: Get today's coach review
      responses:
        "200":
          description: Today's review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachReview"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
      required:
        - id
        - email
    SignupInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      required:
        - id
        - name
        - age
        - gender
        - heightCm
        - currentWeightKg
        - goalWeightKg
        - bodyType
        - goals
        - fitnessLevel
        - gymAccess
        - workoutDaysPerWeek
        - wakeTime
        - sleepTime
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        heightCm:
          type: number
        currentWeightKg:
          type: number
        goalWeightKg:
          type: number
        bodyType:
          type: string
        goals:
          type: array
          items:
            type: string
        targetDate:
          type: ["string", "null"]
        fitnessLevel:
          type: string
        gymAccess:
          type: string
        equipment:
          type: ["string", "null"]
        workoutDaysPerWeek:
          type: integer
        preferredWorkoutTime:
          type: ["string", "null"]
        wakeTime:
          type: string
        sleepTime:
          type: string
        sleepQuality:
          type: integer
        energyLevel:
          type: integer
        stressLevel:
          type: integer
        workSchedule:
          type: ["string", "null"]
        averageDailySteps:
          type: ["integer", "null"]
        allergies:
          type: ["string", "null"]
        dislikedFoods:
          type: ["string", "null"]
        dietStyle:
          type: ["string", "null"]
        foodBudget:
          type: ["string", "null"]
        mealsPerDay:
          type: integer
        waterIntakeLiters:
          type: number
        caffeineUse:
          type: ["string", "null"]
        screenTimeBeforeBed:
          type: ["string", "null"]
        skinConcerns:
          type: array
          items:
            type: string
        digestionConcerns:
          type: array
          items:
            type: string
        biggestStruggle:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserProfileInput:
      type: object
      required:
        - name
        - age
        - gender
        - heightCm
        - currentWeightKg
        - goalWeightKg
        - bodyType
        - goals
        - fitnessLevel
        - gymAccess
        - workoutDaysPerWeek
        - wakeTime
        - sleepTime
      properties:
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        heightCm:
          type: number
        currentWeightKg:
          type: number
        goalWeightKg:
          type: number
        bodyType:
          type: string
        goals:
          type: array
          items:
            type: string
        targetDate:
          type: string
        fitnessLevel:
          type: string
        gymAccess:
          type: string
        equipment:
          type: string
        workoutDaysPerWeek:
          type: integer
        preferredWorkoutTime:
          type: string
        wakeTime:
          type: string
        sleepTime:
          type: string
        sleepQuality:
          type: integer
        energyLevel:
          type: integer
        stressLevel:
          type: integer
        workSchedule:
          type: string
        averageDailySteps:
          type: integer
        allergies:
          type: string
        dislikedFoods:
          type: string
        dietStyle:
          type: string
        foodBudget:
          type: string
        mealsPerDay:
          type: integer
        waterIntakeLiters:
          type: number
        caffeineUse:
          type: string
        screenTimeBeforeBed:
          type: string
        skinConcerns:
          type: array
          items:
            type: string
        digestionConcerns:
          type: array
          items:
            type: string
        biggestStruggle:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        currentWeightKg:
          type: number
        sleepQuality:
          type: integer
        energyLevel:
          type: integer
        stressLevel:
          type: integer
    Plan:
      type: object
      required:
        - id
        - userId
        - goalType
        - calorieTarget
        - proteinTargetG
        - waterTargetL
        - stepsTarget
        - sleepTargetHours
        - weeklyPace
        - workoutSchedule
        - keyHabits
        - coachNotes
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        goalType:
          type: string
        calorieTarget:
          type: integer
        proteinTargetG:
          type: integer
        waterTargetL:
          type: number
        stepsTarget:
          type: integer
        sleepTargetHours:
          type: number
        weeklyPace:
          type: string
        workoutSchedule:
          type: string
        keyHabits:
          type: array
          items:
            type: string
        coachNotes:
          type: string
        warnings:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    JournalEntry:
      type: object
      required:
        - id
        - userId
        - date
        - followedSchedule
        - hitProtein
        - stayedNearCalories
        - workedOut
        - drankWater
        - sleptOnTime
        - energyRating
        - skinBloatingRating
        - biggestWin
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        followedSchedule:
          type: boolean
        hitProtein:
          type: boolean
        stayedNearCalories:
          type: boolean
        workedOut:
          type: boolean
        drankWater:
          type: boolean
        sleptOnTime:
          type: boolean
        energyRating:
          type: integer
        skinBloatingRating:
          type: integer
        whatWentWrong:
          type: ["string", "null"]
        biggestWin:
          type: string
        needHelpWith:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    JournalEntryInput:
      type: object
      required:
        - followedSchedule
        - hitProtein
        - stayedNearCalories
        - workedOut
        - drankWater
        - sleptOnTime
        - energyRating
        - skinBloatingRating
        - biggestWin
      properties:
        followedSchedule:
          type: boolean
        hitProtein:
          type: boolean
        stayedNearCalories:
          type: boolean
        workedOut:
          type: boolean
        drankWater:
          type: boolean
        sleptOnTime:
          type: boolean
        energyRating:
          type: integer
        skinBloatingRating:
          type: integer
        whatWentWrong:
          type: string
        biggestWin:
          type: string
        needHelpWith:
          type: string
    CoachReview:
      type: object
      required:
        - id
        - userId
        - date
        - dailyScore
        - biggestWin
        - biggestMistake
        - whatSlowedProgress
        - exactFixForTomorrow
        - onPace
        - strictCoachMessage
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        dailyScore:
          type: integer
        biggestWin:
          type: string
        biggestMistake:
          type: string
        whatSlowedProgress:
          type: string
        exactFixForTomorrow:
          type: string
        onPace:
          type: boolean
        strictCoachMessage:
          type: string
        createdAt:
          type: string
          format: date-time
...