# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ascend API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupInput {
  email: string;
  /** @minLength 8 */
  password: string;
}